Split up dialog script and FreeRoamTutorial script.
