+++ VR Maze/Assets/Scripts/DialogController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: d369eaf07eb878d469cb94930a18fdd8
+timeCreated: 1491868290
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
