+++ VR Maze/Assets/Scripts/FreeRoamTut.cs
@@ -6,52 +6,26 @@ using UnityEngine.UI;
 
 namespace Assets.Scripts
 {
-	public class FreeRoamTut: MonoBehaviour {
-
-		public Text promptText;
-		public Image promptBox;
+	public class FreeRoamTut: MonoBehaviour{
 
 		private string[] msgState = {"Welcome to Free Roam!\nLook around using your VR headset!","Great!\nTo move forward, look down!","To stop, look up towards the sky.\nYou don't need to look all the way up.","Well done!\nNow, find your way out of this maze."};
 		private int msgIndex = 0;
-		private bool taskCompleted = false;
+		//private bool taskCompleted = false;
+
+		private GameObject player;
+
 
 		void Start(){
-			displayPrompt (false);
-			StartCoroutine (prompt (3.0f));
+			player = GameObject.Find ("MainPlayer");
+
 		}
 
-		void Update(){
-			
-			if (taskCompleted) {
-				msgIndex++;	
-			}
-		}
-			
-		/// <summary>
-		/// IEnumberator works in conjunction with StartCoroutine()
-		/// Creates typewriter effect for displaying dialog.
-		/// </summary>
-		/// <returns>Edits promptBox</returns>
-		IEnumerator prompt(float delay){
-			yield return new WaitForSeconds(delay);
-			displayPrompt (true);
-			for (int i = 0; i < (msgState[msgIndex].Length+1); i++)
-			{
-				promptText.text = msgState[msgIndex].Substring(0, i);
-				yield return new WaitForSeconds(.03f);
-			}
+		public string getMsg(){
+			return msgState [msgIndex];
 		}
 
-		public void displayPrompt(bool status){
-			promptBox.enabled = status;
-			promptText.enabled = status;
-			disableWalking (status);
-		}
-
-		public void disableWalking(bool status){
-			GameObject player = GameObject.Find ("MainPlayer");
-			WalkingScript walkingController = player.GetComponent<WalkingScript> ();
-			walkingController.freezePlayer = status;
+		public string msgOne(){
+			return  msgState [msgIndex];
 		}
 	}
 }
